requirements.txt
Pillow == 9.4.0
matplotlib == 3.7.1
customtkinter == 5.1.2
numpy == 1.24.2